Move it to its own pass
